+++ migrations/0008_client_logout_endpoints.down.sql
@@ -1,4 +1,0 @@
-ALTER TABLE clients
-    DROP COLUMN backchannel_logout_uri,
-    DROP COLUMN frontchannel_logout_uri,
-    DROP COLUMN post_logout_redirect_uris;