+++ migrations/0011_auth_session_mfa_pending.down.sql
@@ -1,2 +1,0 @@
-ALTER TABLE auth_sessions
-    DROP COLUMN password_verified_at;